Designed a plain old c style api. Right now, it only handles the most the common use case: interpreting an fsk-encoded stackmat signal. I plan to add support for interpreting a digital stackmat signal later.
This is a bit lame, since I'm rather proud of our c++ classes, but
nothing like swig exists for javascript and java's jni. With enough work,
we could export our c++ OOP api to js and java, but it would require a
lot of handwritten glue code (see embind.cpp for what this looks like
for javascript).

I plan to get rid of embind.cpp (too much hassle to maintain) and use
this instead. This api will also be used by the android NDK.